+++ opentelemetry-api/src/opentelemetry/metrics/__init__.py
@@ -33,12 +33,33 @@ class DefaultMetricHandle:
     Used when no MetricHandle implementation is available.
     """
 
+    def add(self, value: ValueT) -> None:
+        """No-op implementation of `CounterHandle` add.
+
+        Args:
+            value: The value to add to the handle.
+        """
+
+    def set(self, value: ValueT) -> None:
+        """No-op implementation of `GaugeHandle` set.
+
+        Args:
+            value: The value to set to the handle.
+        """
+
+    def record(self, value: ValueT) -> None:
+        """No-op implementation of `MeasureHandle` record.
+
+        Args:
+            value: The value to record to the handle.
+        """
+
 
 class CounterHandle:
     def add(self, value: ValueT) -> None:
         """Increases the value of the handle by ``value``.
         Args:
-            value: The value to record to the handle.
+            value: The value to add to the handle.
         """
 
 
@@ -46,7 +67,7 @@ class GaugeHandle:
     def set(self, value: ValueT) -> None:
         """Sets the current value of the handle to ``value``.
         Args:
-            value: The value to record to the handle.
+            value: The value to set to the handle.
         """
 
 
@@ -75,12 +96,13 @@ class DefaultLabelSet(LabelSet):
     """
 
 
-class Metric:
+class Metric(abc.ABC):
     """Base class for various types of metrics.
     Metric class that inherit from this class are specialized with the type of
     handle that the metric holds.
     """
 
+    @abc.abstractmethod
     def get_handle(self, label_set: LabelSet) -> "object":
         """Gets a handle, used for repeated-use of metrics instruments.
         Handles are useful to reduce the cost of repeatedly recording a metric
@@ -103,6 +125,30 @@ class DefaultMetric(Metric):
             label_set: `LabelSet` to associate with the returned handle.
         """
         return DefaultMetricHandle()
+
+    def add(self, value: ValueT, label_set: LabelSet) -> None:
+        """No-op implementation of `Counter` add.
+
+        Args:
+            value: The value to add to the counter metric.
+            label_set: `LabelSet` to associate with the returned handle.
+        """
+
+    def set(self, value: ValueT, label_set: LabelSet) -> None:
+        """No-op implementation of `Gauge` set.
+
+        Args:
+            value: The value to set the gauge metric to.
+            label_set: `LabelSet` to associate with the returned handle.
+        """
+
+    def record(self, value: ValueT, label_set: LabelSet) -> None:
+        """No-op implementation of `Measure` record.
+
+        Args:
+            value: The value to record to this measure metric.
+            label_set: `LabelSet` to associate with the returned handle.
+        """
 
 
 class Counter(Metric):
@@ -209,9 +255,9 @@ class Meter:
             label_keys: The keys for the labels with dynamic values.
             enabled: Whether to report the metric by default.
             monotonic: Configure a counter or gauge that accepts only
-            monotonic/non-monotonic updates.
+                monotonic/non-monotonic updates.
             absolute: Configure a measure that does or does not accept negative
-            updates.
+                updates.
         Returns: A new ``metric_type`` metric with values of ``value_type``.
         """
         # pylint: disable=no-self-use
@@ -223,6 +269,34 @@ class Meter:
             labels: A dictionary representing label key to label value pairs.
         Returns: A `LabelSet` object canonicalized using the given input.
         """
+
+
+class DefaultMeter(Meter):
+    """The default Meter used when no Meter implementation is available."""
+
+    def record_batch(
+        self,
+        label_set: LabelSet,
+        record_tuples: Sequence[Tuple["Metric", ValueT]],
+    ) -> None:
+        pass
+
+    def create_metric(
+        self,
+        name: str,
+        description: str,
+        unit: str,
+        value_type: Type[ValueT],
+        metric_type: Type[MetricT],
+        label_keys: Sequence[str] = (),
+        enabled: bool = True,
+        monotonic: bool = False,
+        absolute: bool = True,
+    ) -> "Metric":
+        # pylint: disable=no-self-use
+        return DefaultMetric()
+
+    def get_label_set(self, labels: Dict[str, str]) -> "LabelSet":
         # pylint: disable=no-self-use
         return DefaultLabelSet()
 
@@ -244,7 +318,12 @@ def meter() -> Meter:
 
     if _METER is None:
         # pylint:disable=protected-access
-        _METER = loader._load_impl(Meter, _METER_FACTORY)
+        try:
+            _METER = loader._load_impl(Meter, _METER_FACTORY)  # type: ignore
+        except TypeError:
+            # if we raised an exception trying to instantiate an
+            # abstract class, default to no-op tracer impl
+            _METER = DefaultMeter()
         del _METER_FACTORY
 
     return _METER