+++ ext/opentelemetry-ext-jaeger/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Export span status ([#367](https://github.com/open-telemetry/opentelemetry-python/pull/367))
+- Export span kind ([#387](https://github.com/open-telemetry/opentelemetry-python/pull/387))
+
 ## 0.3a0
 
 Released 2019-12-11